Move storage & import config from config.h to invstorage.h
