+++ src/httpinterface.h
@@ -6,6 +6,7 @@
 #include <curl/curl.h>
 #include "json/json.h"
 
+#include "config.h"
 #include "utils.h"
 
 struct HttpResponse {
