+++ src/sqlstorage.cc
@@ -1,7 +1,10 @@
 #include "sqlstorage.h"
 
+#include <boost/lexical_cast.hpp>
 #include <iostream>
 #include <memory>
+
+#include <boost/lexical_cast.hpp>
 
 #include "logging.h"
 #include "sql_utils.h"
